Fix service discovery
- Don't get involved in mutating responses in the transport, this
doesn't work for chunked responses.
- Transparently proxy everything except the discovery responses.  It's
not likely that we will ever need to modify the content for actual API
responses.  The disovery responses only need to be modified so clients
continue to use the proxy.  (Maybe also to remove auth information, but
it doesn't appear that this is respected by the gcloud SDK tools.)